requirements.txt
requests
pillow
Flask
gunicorn
praw
beautifulsoup4